+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `editLaneTitle` and `onLaneUpdate` props (availability to inline edit lane
   title)
 
+### Fixes
+
+* #201 Fixed Warning: Failed prop type by replacing `react-popover` with
+    `react-popopo`
+
+### Improvements
+
+* Removed `react-popover`'s CSS classes. 
+
 ## [2.2.0-alpha.1] - 2019-07-01
 
 ### Added